Split Log4j plugins and Tomcat components
Log4j Core plugins and Tomcat components need to be in different
classloaders:

 * Log4j core plugins should be in the system classloader,
 * Tomcat components must be in the server/common classloader.
