+++ log4j-tomcat-juli/pom.xml
@@ -21,13 +21,18 @@
     <artifactId>log4j-plugins-parent</artifactId>
     <version>3.0.0-SNAPSHOT</version>
   </parent>
-  <artifactId>log4j-tomcat</artifactId>
-  <name>Tomcat components for Log4j</name>
-  <description>Tomcat 8.5.x/9.x/10.x components to improve support for Log4j API.</description>
+  <artifactId>log4j-tomcat-log4j</artifactId>
+  <name>Log4j Core plugins for Tomcat</name>
+  <description>Log4j Core plugins that improve logging support on Tomcat</description>
   <dependencies>
     <dependency>
+      <groupId>org.apache.logging.log4j</groupId>
+      <artifactId>log4j-core</artifactId>
+      <optional>true</optional>
+    </dependency>
+    <dependency>
       <groupId>org.apache.tomcat</groupId>
-      <artifactId>tomcat-catalina</artifactId>
+      <artifactId>tomcat-juli</artifactId>
       <scope>provided</scope>
     </dependency>
     <dependency>
@@ -43,16 +48,6 @@
     <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-params</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.logging.log4j</groupId>
-      <artifactId>log4j-api</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.logging.log4j</groupId>
-      <artifactId>log4j-core</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
