+++ log4j-tomcat-juli/src/main/java/eu/copernik/log4j/tomcat/juli/TomcatLookup.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.copernik.log4j.tomcat;
+package eu.copernik.log4j.tomcat.juli;
 
 import org.apache.juli.WebappProperties;
 import org.apache.logging.log4j.core.LogEvent;
