+++ log4j-tomcat-juli/src/test/java/eu/copernik/log4j/tomcat/juli/TomcatLookupTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package eu.copernik.log4j.tomcat;
+package eu.copernik.log4j.tomcat.juli;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
