+++ pom.xml
@@ -46,6 +46,7 @@
 
   <modules>
     <module>log4j-tomcat</module>
+    <module>log4j-tomcat-juli</module>
   </modules>
 
   <scm>
